backend/requirements.txt
fastapi
uvicorn[standard]
supabase>=2.0.0
pydantic>=2.0.0
python-dotenv
httpx
python-jose[cryptography]
python-multipart